requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
wikipedia
langchain_huggingface
pandas
langchain-groq
huggingface_hub
langchain-openai
langchain-experimental
neo4j==5.24.0 #  neo4j-5.24.0
tiktoken
yfiles_jupyter_graphs
streamlit 
langchain_chroma
chromadb
pypdf



## pip install --upgrade streamlit
#local packages --------------------------------
-e .